+++ roster/roster_test.go
@@ -6,6 +6,7 @@ package roster_test
 
 import (
 	"context"
+	"encoding/xml"
 	"io"
 	"io/ioutil"
 	"reflect"
@@ -13,8 +14,10 @@ import (
 	"strings"
 	"testing"
 
+	"mellium.im/xmlstream"
 	"mellium.im/xmpp/internal/xmpptest"
 	"mellium.im/xmpp/jid"
+	"mellium.im/xmpp/mux"
 	"mellium.im/xmpp/roster"
 	"mellium.im/xmpp/stanza"
 )
@@ -87,5 +90,55 @@ func TestFetch(t *testing.T) {
 				t.Errorf("Wrong items:\nwant=\n%+v,\ngot=\n%+v", tc.items, items)
 			}
 		})
+	}
+}
+
+func TestReceivePush(t *testing.T) {
+	const itemJID = "nurse@example.com"
+	const x = `<iq xmlns='jabber:client' id='a78b4q6ha463' to='juliet@example.com/chamber' type='set'><query xmlns='jabber:iq:roster'><item jid='` + itemJID + `'/></query></iq>`
+
+	d := xml.NewDecoder(strings.NewReader(x))
+	var b strings.Builder
+	e := xml.NewEncoder(&b)
+
+	called := false
+	h := roster.Handler{
+		Push: func(item roster.Item) error {
+			if item.JID.String() != itemJID {
+				t.Errorf("unexpected JID: want=%q, got=%q", itemJID, item.JID.String())
+			}
+			called = true
+			return nil
+		},
+	}
+
+	tok, err := d.Token()
+	if err != nil {
+		t.Errorf("unexpected error popping start token: %v", err)
+	}
+	start := tok.(xml.StartElement)
+	m := mux.New(roster.Handle(h))
+	err = m.HandleXMPP(struct {
+		xml.TokenReader
+		xmlstream.Encoder
+	}{
+		TokenReader: d,
+		Encoder:     e,
+	}, &start)
+	if err != nil {
+		t.Errorf("unexpected error in handler: %v", err)
+	}
+	err = e.Flush()
+	if err != nil {
+		t.Errorf("unexpected error flushing encoder: %v", err)
+	}
+
+	if !called {
+		t.Errorf("expected push handler to be called")
+	}
+
+	out := b.String()
+	if out != "" {
+		t.Errorf("want=%q, got=%q", "", out)
 	}
 }